requirements.txt
icalendar>=5.0.0
feedparser>=5.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
peewee>=3.17.0
rapidfuzz>=3.6.0
pyyaml>=6.0
python-dotenv>=1.0.0
